restructure repository directories
