+++ .gitignore
@@ -1,2 +1,1 @@
-build.sh
-test-pod.yaml
+ssm-sync
